// src/lib/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      listings: {
        Row: {
          created_at: string
          description: string | null
          id: string
          image_urls: string[] | null
          location: string
          pallet_condition: Database["public"]["Enums"]["pallet_condition"]
          pallet_type: Database["public"]["Enums"]["pallet_type"]
          price: number
          price_per_piece: boolean
          quantity: number
          title: string
          transport_options: Json | null
          user_id: string
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: string
          image_urls?: string[] | null
          location: string
          pallet_condition: Database["public"]["Enums"]["pallet_condition"]
          pallet_type: Database["public"]["Enums"]["pallet_type"]
          price: number
          price_per_piece?: boolean
          quantity: number
          title: string
          transport_options?: Json | null
          user_id: string
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: string
          image_urls?: string[] | null
          location?: string
          pallet_condition?: Database["public"]["Enums"]["pallet_condition"]
          pallet_type?: Database["public"]["Enums"]["pallet_type"]
          price?: number
          price_per_piece?: boolean
          quantity?: number
          title?: string
          transport_options?: Json | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "listings_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      pallet_condition: "UUS" | "KASUTATUD_HELE" | "KASUTATUD_TUME"
      pallet_type: "EUR/EPAL" | "FIN" | "MUU"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never